Refactored task 07. Removed unnecessary things

+++ Course/07/Employee.cs
@@ -22,7 +22,7 @@ namespace Task_07
 
             foreach (Position position in Positions)
             {
-                totalWorkDays += (int)position.GetWorkDays();
+                totalWorkDays += position.GetWorkDays();
             }
 
             return totalWorkDays;
